+++ sugarloaf/src/glyph/section.rs
@@ -1,0 +1,12 @@
+// glyph module code along with comments was originally retired from glyph-brush
+// https://github.com/alexheretic/glyph-brush
+// glyph-brush was originally written Alex Butler (https://github.com/alexheretic)
+// and licensed under Apache-2.0 license.
+
+mod builder;
+mod owned;
+mod refed;
+
+pub use builder::*;
+pub use owned::*;
+pub use refed::*;